refactor: 사용 않는 객체 삭제 및 패키지 이동
