+++ app/src/main/java/com/techcourse/controller/UserController.java
@@ -22,11 +22,11 @@ public class UserController {
         final String account = request.getParameter("account");
         log.debug("user id : {}", account);
 
+
         final ModelAndView modelAndView = new ModelAndView(new JsonView());
         final User user = InMemoryUserRepository.findByAccount(account)
                 .orElseThrow();
 
-        modelAndView.addObject("user", user);
-        return modelAndView;
+        return modelAndView.addObject("user", user);
     }
 }
