+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerMappings.java
@@ -1,6 +1,5 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
